Refactor sequences (#1)
* reorganized data code, and added docs to the sequences

* added docs for the zero filled reconstruction

* added docs to fourier functions

* added docs to the data utils

* removed af8 training scripts

* corrected sequences in af4 training scripts

* adapted torch datasets to new structure of data

* changed datasets name for greater consistency with sequences

* adapted torch scripts to new datasets name and location

* removed the quick fix for torch training with scale factors

* added modules docs to fastmri data handlers

* added module doc to fourier utils
