+++ cascadenet_approach_af4.py
@@ -8,9 +8,8 @@ from keras.utils.vis_utils import model_to_dot
 from keras_tqdm import TQDMCallback
 import tensorflow as tf
 
-from data import MaskedScaled2DSequence
-from evaluate import psnr, ssim
 from cascading import cascade_net
+from fastmri_sequences import Masked2DSequence
 
 
 
@@ -44,8 +43,8 @@ n_volumes_val = 199
 
 # generators
 AF = 4
-train_gen = MaskedScaled2DSequence(train_path, af=AF, inner_slices=8, rand=True, scale_factor=1e6)
-val_gen = MaskedScaled2DSequence(val_path, af=AF, scale_factor=1e6)
+train_gen = Masked2DSequence(train_path, af=AF, inner_slices=8, rand=True, scale_factor=1e6)
+val_gen = Masked2DSequence(val_path, af=AF, scale_factor=1e6)
